+++ src/main/java/org/osiam/resources/scim/Group.java
@@ -108,11 +108,7 @@ public final class Group extends Resource implements Serializable {
         private Set<MemberRef> members = new HashSet<>();
 
         /**
-         * creates a new Group.Builder based on the given displayName and group. All values of the given group will be
-         * copied expect the displayName will be be overridden by the given one
-         *
-         * @param displayName the new displayName of the group
-         * @param group       a existing group
+         * @deprecated Change the display name with {@link #displayName(String)}. Will be removed in 1.12 or 2.0.
          */
         public Builder(String displayName, Group group) {
             super(group);
@@ -159,18 +155,36 @@ public final class Group extends Resource implements Serializable {
             }
         }
 
+        /**
+         * @deprecated You should not need to set the ID with a client. Will be removed in 1.12 or 2.0.
+         */
+        @Deprecated
         @Override
         public Builder setId(String id) {
             super.setId(id);
             return this;
         }
 
+        /**
+         * @deprecated You should not need to set the meta attribute with a client. Will be removed in 1.12 or 2.0.
+         */
+        @Deprecated
         @Override
         public Builder setMeta(Meta meta) {
             super.setMeta(meta);
             return this;
         }
 
+        @Override
+        public Builder externalId(String externalId) {
+            super.externalId(externalId);
+            return this;
+        }
+
+        /**
+         * @deprecated Use {@link #externalId(String)}. Will be removed in 1.12 or 2.0.
+         */
+        @Deprecated
         @Override
         public Builder setExternalId(String externalId) {
             super.setExternalId(externalId);
@@ -185,6 +199,17 @@ public final class Group extends Resource implements Serializable {
         @Deprecated
         public Builder setSchemas(Set<String> schemas) {
             super.setSchemas(schemas);
+            return this;
+        }
+
+        /**
+         * Sets the display name (See {@link Group#getDisplayName()}).
+         *
+         * @param displayName the display name to set
+         * @return the builder itself
+         */
+        public Builder displayName(String displayName) {
+            this.displayName = displayName;
             return this;
         }